Add new subnet doing a compare
Input Cidr range txt file and compare against subnet cidrs already used, pick one and build a new subnet in var az that is not already in use. This is a hack but works well for addressing Cidr. I realize there is the module function !Cidr or Fn::Cidr that can also be used but this was a quick and easily solution to address instead.